Minor code improvements
added some documentation to xcorr_fft and made sure that the
return type is always a numpy array. Cases where the length of the
inputs were zero previously cause errors.

improved the __str__ method of DataCurve. Cases where the length
of the array was zero were not covered by the __str__ method

cover cases in Model.__str__ for cases an iterated object is
not a FittingParameter