Worked on packing app for osx
Now ChiSurf can be packed as an osx app. There
were issues when starting the app by clicking on
the icon with the mouse (see https://github.com/pyinstaller/pyinstaller/issues/3820).

The issues were resolved by compiling the gui entry
point with nuitka and using this as a start point
for the .app