Added typing_extensions package
cover case when typing extension cannot be imported. When the
setup.py is called from an environment that does not have
typing_extensions installed the setup.py fails, even though
typing_extensions is not needed to run setup.py. Thus, the
import of typing in chisurf now imports typing as None if the
library does not exist.