Fixed issue with the dmg-build
The build-dmg uses the ouput of

python -c "import chisurf; import pathlib; print(pathlib.Path(chisurf.__file__).parent.absolute())"

to determine the path of the python library. The warning was outputed to the stdout.
Hence, when the warning is printed to std the above comment does not get the path
correctly.